Discord mod or Guessing game
I am planning on maybe doing a guessing game project with a very complex number, something like 300294291.422.
My other option would be a discord mod that can inform a user about traits and benefits of a certain character, or a weapon in a game. 
This mod would help the user save time learning more about the characters and weapons in the game, and help him/her find the best characters or weapons.